requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2
django-extensions==3.2.3
django-seed==0.2.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
factory_boy==3.3.1
Faker==33.3.0
idna==3.10
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.12.2
urllib3==2.3.0
